ArmPkg: Allow disabling of ArmSoftFloatLib build
BZ: https://bugzilla.tianocore.org/show_bug.cgi?id=2140

This commit adds the define "SOFTFLOAT_ENABLE" that will determine
whether the ArmPkg\ArmSoftFloatLib module is built.
By default it is set to TRUE when the ARCH is ARM, but it can be
turned off in the case of VS builds (non-gcc builds), which fail
to build berkeley_softfloat-3, which is specifically targeted for gcc.

Signed-off-by: Baptiste Gerondeau <baptiste.gerondeau@linaro.org>