Add SOFTFLOAT_ENABLE for ArmSoftFloatLib building
REF:https://bugzilla.tianocore.org/show_bug.cgi?id=2140

This commit adds a define "SOFTFLOAT_ENABLE" that will determine
whether the ArmPkg\ArmSoftFloatLib module is built.
By default it is set to TRUE when the ARCH is ARM, but it can be
turned off in the case of MSFT builds, which fail to build
berkeley_softfloat-3.
To do so, add "-D SOFTFLOAT_ENABLE=FALSE" to your build command.